Raise cMart sale ribbon above the header, show it on cToon info modal
- Move the active-sale ribbon out of the scrollable cToons grid so it
  floats over the top-right corner of the whole cMart panel instead of
  being clipped near the card row.
- Share the active-sale lookup (with its 30s cache) between the cMart
  showcase endpoint and the cToon info modal endpoint, so the modal can
  tell when the viewed cToon is currently on sale.
- When a cToon is on sale, show the sale's ribbon image in the info
  modal's corner badge instead of the new/coming-soon image.
